Move webapp code into webapp folder.
