+++ frontend/webapp/next.config.mjs
