+++ frontend/webapp/src/app/components/AssistantInteraction.js
