+++ frontend/webapp/src/app/components/assistantButton.js
