+++ frontend/webapp/src/app/components/assistantPicker.js
