+++ frontend/webapp/src/app/favicon.ico
