+++ frontend/webapp/src/app/pages/interactor.js
