+++ frontend/webapp/src/app/utils/utils.js
